requirements.txt
streamlit==1.41.1

langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.4

faiss-cpu==1.9.0.post1

pypdf==5.1.0
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3

ollama==0.4.4

python-dotenv==1.0.1
tqdm==4.67.1

rich==13.9.4
